feat(constraints): Add option to hide selected elements when multiple can be selected
The goal is to avoid having the user add multiple times the same value and cause confusion when they remove one of them and realize there are more.

The list can be refreshed after each option selection, but not when an option is removed. To reflect the changes, the user has to change the input.